Rewrite to emphasize standardization
The majority of the changes here are "non-normative", and simply a matter of more clearly emphasizing the way in which the process works. It shifts the focus away from the implementation detail of code sharing, toward the fact that these would, like all web platform APIs, be created through the standards process. In doing this, it refocuses our goals and the benefits involved.

Notably, this renames the proposal from "permafills" to "layered APIs", as "permafills" was giving misleading impressions.

At the same time, this fleshes out some of the examples, also switching from browser:x|y to std:x|y, and removing the stdsrc="" proposal.