Rewrite with a stronger dependency on package name maps
This incorporates some of the ideas touched on in the issue tracker, mostly around using package name maps to provide a better fallback story and web developer control. It touches on the following issues:

* Fixes #10, fixes #14, fixes #15, and fixes #24 by moving away from std:x|y and its attendant problems.
* Opens the door to solutions for #5 based on the speculative package name map fallback ideas.

It also adds a bit more motivation to the intro, and rearranges to emphasize the standards process, before diving into many details on the importing infrastructure.

For now we delete the "Trying these ideas out" section, and spec.md, as they are all about the old infrastructure.